+++ Packages/com.unity.services.mediation/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [0.2.1-preview.1] - 2021-09-17
+
+### Added
+- Test validating gradle version to display a more meaningful error message.
+
+### Changed
+- Overhauled the Ad Network adapters list UI
+
+### Fixed
+- Archived Ad Units will no longer be displayed in the ad units list.
+
 ## [0.2.0-preview.5] - 2021-08-25
 
 ### Added